Turn terminfo_data.jl into a self-updating quine
Is it brilliant or cursed? That's up for debate, but I think it's neat.

This change also revealed some minor discrepancies between the NCurses
source (which we use nwo) and the processed source I originally
used (from some python package? I can't quite recall).

All of the minor discrepancies are cosmetic, with the exception of:
- memory_below moving down by one, and
- backspace_with_bs changing to backspace_if_not_bs

Since we are now using NCurses's source directly, I'm inclined to trust
these changes.