Switch the field ordering of TermCapability
The preferred code to use is the "long" symbol, for readability when
used. I think it makes sense to have it as the first field to reflect
that. While we're at it, we may as well update the field names to better
indicate that "short" is the actual capname, and just call "long"
"name".