Add `Pipe` to the documentation (#53202)
The reasoning is that `Pipe` is specifically documented in other
functions that are part of the public API (e.g. `redirect_stdio()`), so
it should be documented too. Same goes for `link_pipe!()` since it's
mentioned in the `Pipe` docstring.

The other function that could be documented is `open_pipe!()`, but that
specifically applies to `PipeEndpoint`'s and I could imagine that we
want it to be difficult to create a half-initialized `Pipe`
:sweat_smile: I wrote a docstring for `link_pipe!()` based off the libuv
docs: https://docs.libuv.org/en/v1.x/pipe.html#c.uv_pipe