Qualify public, unexported bindings in REPL help (#52524)
Fixes #52472, which was caused by `names` being changed to also return
public, unexported symbols in #50105. Note that this restores previous
behavior. A case could be made to instead add the public, unexported
bindings as suggestions with the appropriate qualification.

Not entirely sure how to test this so I'd welcome any suggestions.

---------

Co-authored-by: Jameson Nash <vtjnash@gmail.com>