document Base.Chars arguments of startswith/endswith (#52741)
Since Julia 0.3 (#6779), the `startswith` and `endswith` functions have
accepted not only a string as the prefix/suffix, but also a character or
a tuple/vector/set of characters. However, this wasn't fully documented.

(The docstring actually mentioned vectors and sets of characters, but
this was inconsistent with the given signature.)

Also updates `Base.Chars` to allow `AbstractSet` types, not just `Set`,
which seems to have been an oversight.