Fix interpreter_exec.jl test (#53218)
This test was supposed to check that we correctly handled PhiNodes in
uninferred code in both the interpreter and the compiler. However, the
compiler path wasn't actually exercised, because the `inferred=true`
part of this forced it to be skipped. Drop that and fix the exposed
issues in the compiler where we didn't handle PhiNodes properly.

---------

Co-authored-by: Shuhei Kadowaki <aviatesk@gmail.com>
Co-authored-by: Jameson Nash <vtjnash@gmail.com>
Co-authored-by: Shuhei Kadowaki <40514306+aviatesk@users.noreply.github.com>