More cautious legacy_color conversion in Logging
While the original line

    color = StyledStrings.Face(foreground=StyledStrings.Legacy.legacy_color(color))

does a perfectly adequate job of mapping old color names to
`SimpleColor`s, it neglects the newer capability of using named `Faces`
with logging.

It is better to first check that the value of `color` can indeed be
recognised as a `legacy_color`-compatible named color, before using it
as the foreground value.