Make zero on array of arrays etc apply recursively (#51458)
I wonder if this breaks things, in practice. It shouldn't. Since old
code behavior errored for the cases I am aware of.
As discussed in #38064, this definition is needed to be consistent with
our other linear algebra operations,
and with us considering a vector of vectors etc as a vector space.

Closes #38064