Set LLVM issue tracker URL to Julia's issue tracker (#53128)
This patch sets the default issue tracker URL for LLVM to Julia's issue
tracker. LLVM issues coming from Julia should probably go through
Julia's issue tracker first and then can be reported upstream if there
is an issue with something in upstream LLVM. Having this set to the
default is moderately confusing for users as it refers to the LLVM issue
tracker, which probably isn't wanted.