Correct annotation calculation with annot substr
The commit e1a76bb40668d80ea84a0c2fc458010bf2a4501e was ultimately
reverted due to it seeming to introduce an abort signal in CI. It turns
out there's an underlying issue with a compiler flag being broken,
however I've also noticed that the annotation calculation for SubStrings
was incorrect, as it didn't take into account the substring offset or
width. We might as well fix this.