Merge duplicate entries in results.json
Tests in loops produce duplicate entries, which isn't very helpful. By
calculating a "result ID" based on the location (file and line
number) of the @test statement and the test status, we can avoid adding
duplicate entries to the results list.

To give an indication of how many times a @test was run, we can store
this information in a result_counts dict and then amend the canonical
result's name with the count.