Fix results.json generation after chunk-splitting
In 90afb7c672b7 the results.json file generated for BuildKite was split
into several using Iterators.partition. Unfortunately, since the
json_repr method for vectors was defined for "Vector" not
"AbstractVector", the "SubArray"s produced by the chunking hit the
fallback method, and so the results.json became one giant
string (oops!).

Widening the json_repr method to apply to "AbstractVector" should be all
that's needed to resolve this.

Reported-by: Elliot Saba <staticfloat@gmail.com>