Remove stale explicit imports from many stdlibs (#53464)
Found via my new https://github.com/ericphanson/ExplicitImports.jl

```julia
using ExplicitImports

ExplicitImports.inspect_session(; inner=print_stale_explicit_imports)
```

These are not being used in the modules they are being imported into (or
in some cases they are being used, but only in a qualified way where the
explicit import is unnecessary). It can happen that someone imports a
name into a module in order to have it available in that namespace, to
then access it from another module (e.g. `TOML.Internals` seems to work
like this). I've tried to notice those situations and not remove the
imports in cases where I suspect the import is playing such a role, but
hopefully CI will figure out if I got it right in all cases.

I only looked at stdlibs that happened to be loaded in my Julia session
and whose code is in this repo, so there are probably more stale
explicit imports out there.