Do not consider inlining cost for precompilation
This helps to minimize the amount of failed precompiles
that occur by eliding the inlining cost check. It is
likely this will lead to slightly larger sys/pkgimg
sizes. However in practice it leads to more extensive
and successful precompilation, which can minimize the number
of JIT events which occur when using Julia.