fix number of chunks (#53413)
The manual claims that `a` is split into `nthreads()` chunks, but this
is not true in general. As it was you could get an error, if `length(a)
< nthreads()`, or a number of chunks larger than `nthreads()`, if
`nthreads()` is smaller than `length(a)` but does not divide it. With
`cld`, on the other hand, you always get at most `nthreads()` chunks.