inference: Don't confuse frames in different interpreters
Diffractor's abstract interpreter sometimes needs to do side queries
using the native interpreter. These are pushed onto the regular
inference callstack in anticipation of a future where compiler plugins
may want to recurse from the native interpreter back into the Diffractor
abstract interpreter. However, this introduced a subtle challenge:
When the native interpreter is looking at a frame that is currently
on the inference stack, it would treat them as the same, incorrectly
merging inference across the two abstract interpreters (which have
different semantics and may not be confused). The caches for the two
abstract interpreters were already different, so once things are
inferred, there's no problem (likely because things were already
inferred on the native interpreter), but if not, this could cause
subtle and hard to debug problems.