Fix bootstrap Base precompile in cross compile configuration
The environment variables here need to have target semantics, so
on Windows the path separator needs to be `;` and the path needs
to be converted to windows before being passed through julia.
Because variable assignment is not ordinarily shell expanded,
we also need an extra `$(shell)` to perform this conversion.
Fixes MSYS2 builds of Julia after #53598.