Fix handling of virtual exit node in PostDomTree
The `dom_edges()` for an exit block in the CFG are empty when computing
the PostDomTree, although the algorithm is supposed to treat this as
being dominated by a "virtual" -1 node (which is already included in the
DFSTree). As a result, we can't rely on the loop below this code to
initialize the semi-dominator correctly for us.