enhance the effectiveness of the test cases introduced in #53300 (#53478)
While experimenting with precompilation for external absints on builds
just after #53300 was merged, I found that the test case for
`CustomAbstractInterpreterCaching2.jl` fails if the test case for
`CustomAbstractInterpreterCaching1.jl` isn't run in the same session
beforehand. That is probably because of the previous lack of support for
proper `CodeInstance` caching. To address this, I've changed the tests
to run in separate processes in this commit. Note that it appears that a
recent refactor concerning `CodeInstance` might have resolved this
issue, so the new test cases runs successfully on master. However, I
suspect the fix hasn't been applied to v1.11 yet, we would need more
research.