compiler: Refactor `concrete_eval_invoke`
This passes slightly more information into this function (the full `inst`
rather than just the `stmt`) in order to allow external absint to
access additional fields (the flags and the info) if necessary to make
concrete evaluation decisions. It also splits out the actual
concrete evaluation from the part that just maps the `inst` to
a CodeInstance.