Remove fl_julia_current_{file, line}
These flisp accessor functions make use of the `jl_filename` and
`jl_lineno` globals. I was looking at removing these globals for
unrelated reasons, when I saw that one of the primary uses was
in these flisp accessor, which appear entirely unnecessary.
They are only used to provide a default for an error message,
but the place that puts the error message into a list to return
to julia does actually already have access to this information,
so there's no need for these to look at the globals.

While we're at it, also add a test for this code path, which was
otherwise unexercised in our test suite.